+++ src/main/java/com/nmc/itschool/repository/ScheduleRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ScheduleRepository extends JpaRepository<ScheduleEntity, Integer>{
+public interface ScheduleRepository extends JpaRepository<ScheduleEntity, Long>{
 
     Optional<List<ScheduleEntity>> findAllByAccountId(Integer accountId);
 }
