+++ src/main/java/com/nmc/itschool/repository/SubjectCollectionParentRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface SubjectCollectionParentRepository extends JpaRepository<SubjectCollectionParentEntity, Integer>{
+public interface SubjectCollectionParentRepository extends JpaRepository<SubjectCollectionParentEntity, Long>{
 
     @Query(value = "SELECT * FROM subject_collection_parent_tbl WHERE delete_flag = false" , nativeQuery = true)
     Optional<List<SubjectCollectionParentEntity>> getAll();
